# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Timeseries API API",
    "summary": "A statistical time series analysis for financial and econometric modeling with REST and GraphQL APIs",
    "description": "Econometric time series modeling API with ARIMA and GARCH capabilities. Includes GraphQL interface.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/generate_data": {
      "post": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Generate synthetic time series data",
        "description": "Generate synthetic time series data based on input parameters.\n\nThis endpoint creates synthetic price series for multiple symbols over a specified date range.\nEach symbol starts from its anchor price and follows a random walk with drift.\n\nThe response provides a list of records, each containing a date and values for all symbols.",
        "operationId": "generate_data_endpoint_api_v1_generate_data_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataGenerationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully generated time series data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSeriesDataResponse"
                },
                "example": {
                  "data": [
                    {
                      "date": "2023-01-01",
                      "GME": 150.0,
                      "BYND": 200.0
                    },
                    {
                      "date": "2023-01-02",
                      "GME": 152.3,
                      "BYND": 198.7
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid date format or other input parameters"
          },
          "500": {
            "description": "Internal Server Error - Failed to generate time series data"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fetch_market_data": {
      "post": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Fetch real market data from external sources",
        "description": "Fetch real market data from external sources like Yahoo Finance.\n\nThis endpoint retrieves historical price data for specified symbols over a date range.\nData is obtained from Yahoo Finance via the yfinance library.\n\nThe interval parameter controls the frequency of the data (daily, weekly, monthly).",
        "operationId": "fetch_market_data_endpoint_api_v1_fetch_market_data_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketDataInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully fetched market data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSeriesDataResponse"
                },
                "example": {
                  "data": [
                    {
                      "date": "2023-01-01",
                      "AAPL": 150.0,
                      "MSFT": 250.0
                    },
                    {
                      "date": "2023-01-02",
                      "AAPL": 152.3,
                      "MSFT": 248.7
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid symbols or date range"
          },
          "500": {
            "description": "Internal Server Error - Failed to fetch data from external source"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scale_data": {
      "post": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Scale time series data",
        "description": "Scale time series data using the specified method.\n\nThis endpoint takes raw price data and applies a scaling transformation.\nSupported methods include:\n- standardize: Transforms data to have mean=0 and standard deviation=1\n- minmax: Scales data to a range between 0 and 1\n\nThe input data must contain 'date' and 'price' columns.",
        "operationId": "scale_data_endpoint_api_v1_scale_data_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScalingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully scaled time series data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSeriesDataResponse"
                },
                "example": {
                  "data": {
                    "0": {
                      "date": "2023-01-01",
                      "price": 0.0
                    },
                    "1": {
                      "date": "2023-01-02",
                      "price": 0.5
                    },
                    "2": {
                      "date": "2023-01-03",
                      "price": -0.3
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid data format or missing required columns"
          },
          "500": {
            "description": "Internal Server Error - Failed to scale data"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test_stationarity": {
      "post": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Test for stationarity",
        "description": "Test time series data for stationarity using the Augmented Dickey-Fuller test.\n\nStationarity is a key property for time series analysis, indicating that statistical\nproperties like mean, variance, and autocorrelation are constant over time.\n\nThe test returns:\n- ADF statistic: More negative values suggest stationarity\n- p-value: Smaller values suggest stationarity\n- Critical values: Threshold values at different significance levels\n- is_stationary: Boolean indication based on p-value threshold\n- interpretation: Human-readable explanation of the results\n\nA p-value less than the threshold (default 0.05) indicates stationarity.",
        "operationId": "test_stationarity_endpoint_api_v1_test_stationarity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StationarityTestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully tested time series for stationarity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationarityTestResponse"
                },
                "example": {
                  "adf_statistic": -3.45,
                  "p_value": 0.032,
                  "critical_values": {
                    "1%": -3.75,
                    "5%": -3.0,
                    "10%": -2.63
                  },
                  "is_stationary": true,
                  "interpretation": "The series is stationary (p-value: 0.0320). This means the statistical properties like mean and variance remain constant over time, making it suitable for time series modeling."
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid data format or insufficient data points"
          },
          "500": {
            "description": "Internal Server Error - Failed to run stationarity test"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/price_to_returns": {
      "post": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Convert price data to log returns",
        "description": "Convert price time series data to log returns.\n\nThis endpoint takes price data and calculates the log returns,\nwhich are typically more suitable for statistical modeling.\n\nReturns data represents the percentage change between consecutive price observations.",
        "operationId": "price_to_returns_endpoint_api_v1_price_to_returns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Input Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSeriesDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scale_for_garch": {
      "post": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Scale data for GARCH modeling",
        "description": "Scale time series data specifically for GARCH modeling.\n\nThis endpoint takes return data and applies the appropriate scaling\nfor GARCH volatility modeling.",
        "operationId": "scale_for_garch_endpoint_api_v1_scale_for_garch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Input Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSeriesDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fetch_stooq_data": {
      "post": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Fetch real market data from Stooq",
        "description": "Fetch market data from Stooq data service.\n\nThis endpoint fetches real market data for the specified symbols and date range\nfrom the Stooq data service.",
        "operationId": "fetch_stooq_data_endpoint_api_v1_fetch_stooq_data_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketDataInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully fetched market data from Stooq",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSeriesDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid parameters"
          },
          "500": {
            "description": "Internal Server Error - Failed to fetch market data"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataGenerationInput": {
        "properties": {
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "anchor_prices": {
            "type": "object",
            "title": "Anchor Prices"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "anchor_prices"
        ],
        "title": "DataGenerationInput",
        "description": "Input model for data generation endpoint."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MarketDataInput": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "interval": {
            "type": "string",
            "title": "Interval",
            "default": "1d"
          }
        },
        "type": "object",
        "required": [
          "symbols",
          "start_date",
          "end_date"
        ],
        "title": "MarketDataInput",
        "description": "Input model for market data fetch endpoint."
      },
...
      "ScalingInput": {
        "properties": {
          "method": {
            "type": "string",
            "title": "Method"
          },
          "data": {
            "items": {},
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "method",
          "data"
        ],
        "title": "ScalingInput",
        "description": "Input model for data scaling endpoint."
      },
...
      "StationarityTestInput": {
        "properties": {
          "data": {
            "items": {},
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StationarityTestInput",
        "description": "Input model for stationarity test endpoint."
      },
      "StationarityTestResponse": {
        "properties": {
          "all_symbols_stationarity": {
            "additionalProperties": {
              "type": "object"
            },
            "type": "object",
            "title": "All Symbols Stationarity",
            "description": "Stationarity test results for each symbol"
          },
          "series_stats": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": {
                    "type": "number"
                  },
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Stats",
            "description": "Comprehensive statistics for each series (mean, std, skew, kurtosis)"
          }
        },
        "type": "object",
        "required": [
          "all_symbols_stationarity"
        ],
        "title": "StationarityTestResponse",
        "description": "Response model for stationarity test results."
      },
      "TimeSeriesDataResponse": {
        "properties": {
          "data": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Data",
            "description": "Time series data as a list of records with date and symbol values"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "TimeSeriesDataResponse",
        "description": "Response model for time series data endpoints."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}